+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Use tree-sitter if the parser for Haskell is installed,
+  without requiring the nvim-treesitter plugin.
+
 ## [1.1.0] - 2023-11-05
 
 ### Added