+++ CHANGELOG.md
@@ -6,10 +6,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.2.0] - 2023-11-12
 
 ### Added
 
+- Pragmas: Add all language extensions as choice options [[#12](https://github.com/mrcjkb/haskell-snippets.nvim/pull/12#event-10934080925)].
+  Thanks [@gregorias](https://github.com/gregorias)!
 - Health checks.
 - Use tree-sitter if the parser for Haskell is installed,
   without requiring the nvim-treesitter plugin.
@@ -19,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `no-deprecations` choices [[#11](https://github.com/mrcjkb/haskell-snippets.nvim/pull/11)].
+  Thanks [@gregorias](https://github.com/gregorias)!
 
 ### Fixed